+++ paymentsheet/src/main/java/com/stripe/android/paymentsheet/utils/SelectionUtils.kt
@@ -3,6 +3,7 @@
 package com.stripe.android.paymentsheet.utils
 
 import androidx.annotation.RestrictTo
+import com.stripe.android.model.PaymentMethodExtraParams
 import com.stripe.android.paymentsheet.model.PaymentSelection
 import com.stripe.android.paymentsheet.state.PaymentElementLoader
 
@@ -16,6 +17,20 @@ internal fun PaymentSelection.New.canSave(
         is PaymentElementLoader.InitializationMode.SetupIntent -> true
         is PaymentElementLoader.InitializationMode.DeferredIntent -> {
             initializationMode.intentConfiguration.mode.setupFutureUse != null || requestedToSave
+        }
+    }
+}
+
+internal fun PaymentSelection.getSetAsDefaultPaymentMethodFromPaymentSelection(): Boolean? {
+    return when (this) {
+        is PaymentSelection.New.Card -> {
+            (this.paymentMethodExtraParams as? PaymentMethodExtraParams.Card)?.setAsDefault
+        }
+        is PaymentSelection.New.USBankAccount -> {
+            (this.paymentMethodExtraParams as? PaymentMethodExtraParams.USBankAccount)?.setAsDefault
+        }
+        else -> {
+            null
         }
     }
 }
